+++ config/test.exs
@@ -5,25 +5,7 @@ config :logger,
        backends: []
 
 config :express,
-       apns: [
-         mode: :dev,
-         cert_path: Path.expand("test/fixtures/test_apns_cert.pem"),
-         key_path: Path.expand("test/fixtures/test_apns_key.pem"),
-         key_id: "key_id",
-         team_id: "team_id",
-         auth_key_path: Path.expand("test/fixtures/test_auth_key.p8")
-       ],
-       fcm: [
-         api_key: "your_api_key",
-         collapse_key: "your_collapse_key"
-       ],
+       module: Express.Configuration.Test,
        buffer: [
-        adders_pool_config: [
-          {:name, {:local, :buffer_adders_pool}},
-          {:worker_module, Express.PushRequests.Adder},
-          {:size, 10},
-          {:max_overflow, 2}
-        ],
-        consumers_count: 10,
-        max_size: 10_000
+         max_size: 1000
        ]